requirements.txt
torch==1.10.1
torchvision==0.11.2
torchaudio==0.10.1
cudatoolkit==11.1
mmcv-full==1.4.4
mmsegmentation==0.24.0
